+++ lua/options.lua
@@ -1,6 +1,15 @@
 -- lua/options.lua
 
 --[[
+    Unused language providers
+    Disabling avoids :checkhealth warnings for runtimes not installed.
+--]]
+vim.g.loaded_python3_provider = 0
+vim.g.loaded_ruby_provider    = 0
+vim.g.loaded_perl_provider    = 0
+vim.g.loaded_node_provider    = 0
+
+--[[
     Aspect
 --]]
 -- Global